mcp_project/pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "mcp>=1.9.2",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "nest-asyncio>=1.6.0",
]
